+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## HEAD
 
+* Correctly handle blank username in password reset.
 * Allow configuration of cookie domain.
 * Pass controller to signed_{in,out}_url to allow routes with options/parameters.
 * Fix bug where `signed_in_url` config setting was overwritten.